Add ObjectTreeNodeType
The equality comparer wasn't working for two empty lists, and the
original version of the node class didn't have any way to distinguish
between an empty list and an empty object or primitive value without
using reflection on the type of the value.  So I added a new "node type"
concept that allows that distinguishing, and I used it to fix the
problem with the equality comparisons.